extract FixPieces admin tool
